Added Location.getColony() to get the Colony this Location is associated with or null if not in a Colony().
Refactored all occurances of Unit.getTile().getColony() to use this new Method.

This change was necessary to fix the problems of teachers leaving a colony, joining another one and still retain their teaching progress. This issue is fixed and tested in this commit.

The patch was written by Eric Boumaour.