Add generalized production element, step 1.
This commit adds a new generalized production element, which allows
for production types with several input and output goods, and also
subsumes the basic production and factory production features of
building types. The production element is added to tile types and
building types, but the original semantics are preserved as far as
possible. They will be replaced in subsequent commits.