challenge mod: Roads no longer add +2 furs
Rationale here is that while roads improve access to fur, the incursion
of civilization will have a negative effect on animals, thus should
offset it.  This won't much affect fur availability early in the game,
but should greatly help the oversupply that comes later on after the
road network gets built.