Stop overwriting the JDialog location with a wrong value, fixing BR#2803.
FreeColDialog sets the location correctly in center of screen already.
chooseLocation is incompatible with JDialogs, as it is using a local
coordinate system inside the frame, but a JDialog is a free Window
requiring global window system coordinates.
The local coordinates were causing the dialogs to always appear on
primary screen, as the primary screen got 0,0 in upper left.