Allow adding alternative resource files with images in different resolutions.
Intended use is adding images in doubled size to display on zoomed-in map,
but could be used for something else.
The normal resource key informs the resource of the normal size and is
therefore mandatory, alternative keys with ".a"+number suffixes are optional.
Values of alternative keys should not refer to other resources,
only file names.
The image in most fitting size is automatically choosen and possibly resized.