Fix for bug #2899
This bug is intended to to fix bug #2899 Unnecessary Scrolling after Battle, which appears when multiple messages are displayed by InformationPanel.java. Scrolling may still occur if the number of messages goes above two, but the spacing has been improved dramatically so that each set of text is grouped together better.