Improve OSUtils implementation and abstraction
In order to make the OSUtils class more robust, I have eliminated the operating system input to only the method that requires it. I have also implemented use of the Desktop API (from Java 1.6+) which should work more universally. If that doesn't work, the system then falls back to the old browser launch method.

All errors are logged if thrown. The Desktop API will throw errors as well if the Desktop is not available and then the old method kicks in.