Cleanup logger call not guarded by log condition
Reasoning (credit: IntelliJ IDEA)

Reports logging calls with non-constant arguments which are not surrounded by a guard condition. The evaluation of the arguments of a logging call can be expensive. Surrounding a logging call with a guard clause prevents that cost, when the logging is disabled for the level used by the logging statement. This is especially useful for the least serious level (trace, debug, finest) of logging calls, because those are most often disabled in a production environment.