Improves AI colony development and expansion.
This change has been tested with five games with 300 AI-only turns each. The simple fix was simply stop making tools in every colony. Prior to this commit, tools was prioritized above hammers (slowing down the construction of buildings) and even made elder statesmen work in the fields instead of producing liberty bells.

There are still lots of silly decisions, but these decisions are more complicated to fix -- and will therefor have to wait for the AI colony management rewrite.