Bartelome enables Converts to be FreeColonists
Currently in FreeCol (and Col1), Bartelome joining congress causes native converts to be converted to Free Colonists. The effect is one-time.
This change allows Native Converts to become Free Colonists by being "cleared" from the "clear specialty" menu option.
This change also allows Native Converts to be educated in the outdoor skills that they are skilled at in half the time it would take a Free Colonist to be educated with that skill.