Adding a new AI ColonyPlan that will eventually replace the old one and logic that is currently inside AIColony.
So far, this implementation only includes defining what's the perfect colony for a given location and a debug utility to watch the result. Later, we will be adding a class for refining this plan with the current available workers, buildings and production multipliers -- so that it can be used by AIColony.

The results so far will be verified before improving the committed code, writing tests for edge cases and continuing the work with the AI colony management code.